Move parser package to org.corpus_tools.atomic.console.parser
