+++ plugins/de.uni_jena.iaa.linktype.atomic.atomical/src/main/antlr4/org/corpus_tools/atomic/console/parser/ConsoleCommand.g4
