+++ plugins/de.uni_jena.iaa.linktype.atomic.atomical/src/main/java/de/uni_jena/iaa/linktype/atomic/atomical/console/AtomicalConsole.java
@@ -14,8 +14,8 @@ import org.antlr.v4.runtime.RecognitionException;
 import org.antlr.v4.runtime.Recognizer;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
 import org.corpus_tools.atomic.api.editors.DocumentGraphEditor;
-import org.corpus_tools.atomic.console.ConsoleCommandLexer;
-import org.corpus_tools.atomic.console.ConsoleCommandParser;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandLexer;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.swt.widgets.Display;
