+++ plugins/de.uni_jena.iaa.linktype.atomic.atomical/src/main/java/de/uni_jena/iaa/linktype/atomic/atomical/console/CommandExecutor.java
@@ -7,15 +7,15 @@ import java.util.List;
 import org.antlr.v4.runtime.Token;
 import org.corpus_tools.atomic.api.editors.DocumentGraphEditor;
 import org.corpus_tools.atomic.api.editors.SaltGraphUpdatable;
-import org.corpus_tools.atomic.console.ConsoleCommandBaseListener;
-import org.corpus_tools.atomic.console.ConsoleCommandParser;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.AddOrDeleteEdgeCommandContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.Anno_argsContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.AnnotateCommandContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.DeleteElementCommandContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.HelpCommandContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.NewSpanNodeCommandContext;
-import org.corpus_tools.atomic.console.ConsoleCommandParser.NewStructureNodeCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandBaseListener;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.AddOrDeleteEdgeCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.Anno_argsContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.AnnotateCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.DeleteElementCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.HelpCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.NewSpanNodeCommandContext;
+import org.corpus_tools.atomic.console.parser.ConsoleCommandParser.NewStructureNodeCommandContext;
 import org.corpus_tools.salt.SALT_TYPE;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.corpus_tools.salt.common.SSpan;
