+++ plugins/de.uni_jena.iaa.linktype.atomic.atomical/src/main/java/org/corpus_tools/atomic/console/parser/ConsoleCommand.tokens
