+++ plugins/de.uni_jena.iaa.linktype.atomic.atomical/src/main/java/org/corpus_tools/atomic/console/parser/ConsoleCommandListener.java
@@ -1,5 +1,5 @@
-// Generated from org/corpus_tools/atomic/console/ConsoleCommand.g4 by ANTLR 4.6
-package org.corpus_tools.atomic.console;
+// Generated from org/corpus_tools/atomic/console/parser/ConsoleCommand.g4 by ANTLR 4.6
+package org.corpus_tools.atomic.console.parser;
 import org.antlr.v4.runtime.tree.ParseTreeListener;
 
 /**
